### 2021-03-04 #3085 @whedon
[PRE REVIEW]: ppdiag: Diagnostic Tools for Temporal Point Processes
**Submitting author:** @OwenWard (<a href="http://orcid.org/0000-0002-9678-3542">Owen G. Ward</a>)
**Repository:** <a href="https://github.com/OwenWard/ppdiag" target ="_blank">https://github.com/OwenWard/ppdiag</a>
**Version:** v0.1.0
