### 2021-03-04 OwenWard/ppdiag#52 @mikldk
gitignore
Some files should be removed from the repository, e.g. `.DS_Store` and `.Rapp.history`. See e.g. <https://github.com/github/gitignore/blob/master/R.gitignore>.
