### 2021-03-04 OwenWard/ppdiag#53 @mikldk
README: MMHP 
In README, please explicitely state that MMHP is short for Markov Modulated Hawkes Process. Maybe call section "Markov Modulated Hawkes Process" and mention MMHP afterwards in brackets.
