### 2021-03-04 #3085 @mikldk
@author - thank you for the submission.

1. I created issues at OwenWard/ppdiag#52 and OwenWard/ppdiag#53.
2. I see not all authors contributed with commits to the Github repo. Please confirm that the author list is accurate.
3. Do you have any suggestions for reviewers ([here](https://bit.ly/joss-reviewers) is a list of potential reviewers)?
