### 2021-03-04 #3085 @OwenWard
@mikldk thanks for agreeing to be the editor, appreciate it!

1. Thanks for pointing these out. I'll fix them shortly.
2. The author list is accurate. All authors contributed, either before this project became a github repo or in substantial non code contributions.
3. Some names from that list that might be suitable are alexpghayes, trashbirdecology and  wjakethompson
