### 2021-03-04 edca275
Fixes 52, 53 and update pkgdown with cran link
