+++ README.md
@@ -54,14 +54,17 @@ library(ppdiag)
 hp_obj <- pp_hp(lambda0 = 0.2, alpha = 0.35, beta = 0.8)
 sim_hp <- pp_simulate(hp_obj, end = 200)
 sim_hp
-#>  [1]   7.581703   7.998168   8.614021   8.941138  10.648613  15.364071
-#>  [7]  17.800640  25.491057  28.127978  32.123344  36.650980  40.393909
-#> [13]  46.739837  47.395302  50.294564  62.837118  65.235044  66.340397
-#> [19]  71.383948  77.309497  91.678156  91.905564  95.553557 102.090414
-#> [25] 108.745371 120.906119 122.660710 123.562649 126.055677 149.457071
-#> [31] 156.214770 156.991597 162.095315 162.099438 167.096198 167.398154
-#> [37] 171.239529 174.824605 174.964051 175.205379 175.454114 175.808449
-#> [43] 177.648504 179.547997 187.674615 196.690195 198.482319
+#>  [1]   0.9028774   2.1440211  12.5023753  13.1117594  15.2125103  16.4468785
+#>  [7]  22.0673950  28.0077200  28.7309437  28.8399044  29.8898289  31.2596296
+#> [13]  32.9789068  33.0952675  47.3363303  47.4550222  52.2806143  52.7807790
+#> [19]  53.8471582  70.6502357  71.2346665  71.9704604  74.2900289  77.9300598
+#> [25]  86.2557153  86.9531752  92.3820592 102.7484174 103.2037018 103.2931602
+#> [31] 103.6600985 106.2132004 110.6646722 114.5946863 116.0299836 116.1304621
+#> [37] 116.2030573 116.2219153 116.5640912 116.5834268 116.7923988 117.1433428
+#> [43] 117.3212496 118.2355304 122.7278904 123.5293785 123.9301859 126.7675328
+#> [49] 136.4601124 162.4423253 162.5968625 162.9340964 165.1812875 176.5825227
+#> [55] 180.7287522 185.5611023 185.7220408 186.5938809 190.1311457 192.9218856
+#> [61] 193.2152954 196.3592700
 ```
 
 We can readily evaluate the fit of a homogeneous Poisson process to this
@@ -71,17 +74,20 @@ data.
 est_hpp <- fithpp(sim_hp)
 est_hpp
 #> $lambda
-#> [1] 0.2367969
+#> [1] 0.3157478
 #> 
 #> $events
-#>  [1]   7.581703   7.998168   8.614021   8.941138  10.648613  15.364071
-#>  [7]  17.800640  25.491057  28.127978  32.123344  36.650980  40.393909
-#> [13]  46.739837  47.395302  50.294564  62.837118  65.235044  66.340397
-#> [19]  71.383948  77.309497  91.678156  91.905564  95.553557 102.090414
-#> [25] 108.745371 120.906119 122.660710 123.562649 126.055677 149.457071
-#> [31] 156.214770 156.991597 162.095315 162.099438 167.096198 167.398154
-#> [37] 171.239529 174.824605 174.964051 175.205379 175.454114 175.808449
-#> [43] 177.648504 179.547997 187.674615 196.690195 198.482319
+#>  [1]   0.9028774   2.1440211  12.5023753  13.1117594  15.2125103  16.4468785
+#>  [7]  22.0673950  28.0077200  28.7309437  28.8399044  29.8898289  31.2596296
+#> [13]  32.9789068  33.0952675  47.3363303  47.4550222  52.2806143  52.7807790
+#> [19]  53.8471582  70.6502357  71.2346665  71.9704604  74.2900289  77.9300598
+#> [25]  86.2557153  86.9531752  92.3820592 102.7484174 103.2037018 103.2931602
[…]
