+++ DESCRIPTION
@@ -35,6 +35,8 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.1
+URL: https://owenward.github.io/ppdiag/
+BugReports: https://github.com/OwenWard/ppdiag/issues
 Depends:
     R (>= 3.5)
 Imports: 
