+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: 9c95f00d2505ddd83c4722e019202370715d3a3d
 articles:
   fitting_markov_modulated: fitting_markov_modulated.html
   ppdiag: ppdiag.html
-last_built: 2021-03-04T13:42Z
+last_built: 2021-03-04T23:42Z
 urls:
   reference: https://owenward.github.io/ppdiag/reference
   article: https://owenward.github.io/ppdiag/articles
