### 2021-04-13 OwenWard/ppdiag#54 @wjakethompson
Contributing guidelines
This is part of the JOSS review (openjournals/joss-reviews#3133).

You might consider adding additional guidelines for contributing to the package, including a code of conduct. Generally, I point people to the [`CODE_OF_CONDUCT`](https://github.com/tidyverse/ggplot2/blob/master/CODE_OF_CONDUCT.md) and [`CONTRIBUTING`](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md) files in the [tidyverse/ggplot2](https://github.com/tidyverse/ggplot2) repo as good examples. You probably don't need contributing guidelines as detailed as what is there, but I think it would be helpful to be more clear about the process you would like others to follow, beyond what's currently in the README.
