### 2021-04-13 98d940f
Tweak citation
This is part of openjournals/joss-reviews#3133

Minor tweak to the double citation to match the preferred [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#Citation_Syntax).
