### 2021-04-13 #3133 @wjakethompson
I've completed my initial review. Just few minor comments/suggestions.

* __Community guidelines__: See OwenWard/ppdiag#54
* __References__: See OwenWard/ppdiag#55

Some other thoughts:

* There are authors on the paper that haven't contributed to the repo, but I went ahead and checked that box based on [this comment](https://github.com/openjournals/joss-reviews/issues/3085#issuecomment-791029381).
* I would generally prefer more test coverage than the 56% the package currently has. However, this isn't an acceptance blocker for me, because it looks like the core functionality is tested, and it looks like the non-tested files are generally visualization helpers and wrappers around other functions that are tested (e.g., `drawUniMMHPIntensity.R` has no coverage, but it appears that the main purpose of this function is to wrap `drawHPIntensity.R`, which does have some coverage).
* Finally, there is no explicit mention in the statement of need of the target audience, but it seems fairly clear to me to that the target audience would be researchers using temporal point processes, but not sure if it should be more specific or explicit.
