### 2021-04-26 #3133 @alexpghayes
Also just finished my review. On the whole seems like a well written and useful package. I agree with @wjakethompson that a `CONTRIBUTING.md` file would be helpful.

Other minor feedback:

- It would be nice to provide `print()` methods for `hp` objects, etc
- `pp_diag()` is currently an odd mixture of returning information, and causing side effects (printing as at https://github.com/OwenWard/ppdiag/blob/main/R/pp_diag.R#L46) and plotting. This pattern can be hard for future work to extend upon. I recommend either: (1) creating a diagnostics object that users can interact with programmatically, or (2) returning something like `NULL` and only causing side effects in `pp_diag()`.
- I'm pretty sure that if you use `@keywords internal` rather than `@keywords Internal` you won't need to use `@noRd` when documenting internal functions. Current use is I believe redundant (see https://github.com/OwenWard/ppdiag/blob/main/R/negloglik.R#L10-L11 for example)
- Not that it really matters, but running `styler` once on the package to clean up whitespace, and removing `.DS_Store` files (and adding `.DS_Store` to `.gitignore`) would be minor stylistic improvements.
