### 2021-04-26 #3133 @wjakethompson
> @alexpghayes, @wjakethompson, can you please give a brief status of your reviews? This is merely to give me an impression of the progress and time-frame.

I believe my review is complete. There is still one outstanding PR (OwenWard/ppdiag#55) and one open issue (OwenWard/ppdiag#54), which @OwenWard has indicated the team will be addressing soon. Once those are closed I will check those boxes in my checklist. Or I can go ahead and check them now if that's the preferred process.
