+++ NAMESPACE
@@ -34,6 +34,7 @@ export(pp_hpp)
 export(pp_ksplot)
 export(pp_mmhp)
 export(pp_mmpp)
+export(pp_print)
 export(pp_qqexp)
 export(pp_residual)
 export(pp_simulate)
