+++ R/rawresidual.R
@@ -9,7 +9,7 @@
 #' @param end end of observation period (default final event)
 #' @param steps number of steps for numeric integration (if needed)
 #' @return the raw residual
-#' @noRd
+#' @keywords internal
 #' @examples 
 #' Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
 #' x <- pp_mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, 
