### 2021-05-02 5320f92
change @keywards Internal to internal, also added return NULL for pp_diag
