+++ R/rawresidual.R
@@ -9,13 +9,14 @@
 #' @param end end of observation period (default final event)
 #' @param steps number of steps for numeric integration (if needed)
 #' @return the raw residual
-#' @noRd
+#' @keywords internal
+#' @export
 #' @examples 
 #' Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
 #' x <- pp_mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, 
 #' lambda1 = 1.1, alpha = 0.8, beta = 1.2)
 #' y <- pp_simulate(x, n = 10)
-#' pearsonresidual(x, events = y$events[-1])
+#' ppdiag:::rawresidual(x, events = y$events[-1])
 
 
 
@@ -26,13 +27,15 @@ rawresidual <- function(object, events, start, end, steps = 1000) {
   UseMethod("rawresidual")
 }
 
-
+#' @keywords internal
+#' @export
 rawresidual.default <- function(object, events, start = 0,
                                 end = max(events), steps) {
   cat("Please input the right model. Select from hp, hpp and mmhp.")
 }
 
-
+#' @keywords internal
+#' @export
 rawresidual.hp <- function(object, events, start = 0, 
                            end = max(events), steps = 1000) {
   if(end != max(events)) {
@@ -55,7 +58,8 @@ rawresidual.hp <- function(object, events, start = 0,
   return(N - result)
 }
 
-
+#' @keywords internal
+#' @export
 rawresidual.mmhp <- function(object, events, start = 0,
                              end = max(events), steps = 1000) {
   if(end != max(events)) {
@@ -76,7 +80,8 @@ rawresidual.mmhp <- function(object, events, start = 0,
   return(N - all_Lambda)
 }
 
-
+#' @keywords internal
+#' @export
 rawresidual.hpp <- function(object, events, start = 0, 
                             end = max(events), steps = 1000) {
   if(events[1] == 0){
@@ -91,7 +96,8 @@ rawresidual.hpp <- function(object, events, start = 0,
   return(N - all_Lambda)
 }
 
-
+#' @keywords internal
+#' @export
 rawresidual.mmpp <- function(object, events, start = 0,
                              end = max(events), steps = 1000) {
   if(end != max(events)) {
