+++ R/mmpp_event_state.R
@@ -10,14 +10,13 @@
 #'
 #' @return probability of being in state 1 (active state) at each event,
 #' along with most likely state
-#' @noRd
-#'
+#' @keywords internal
 #' @examples
 #' Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
 #' mmpp_obj <- pp_mmpp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9,
 #'  c = 1.1)
 #'  ## evaluate at some fake event times
-#'  mmpp_event_state(params = mmpp_obj, events = c(1, 2, 3, 5)) 
+#'  ppdiag:::mmpp_event_state(params = mmpp_obj, events = c(1, 2, 3, 5)) 
 mmpp_event_state <- function(params = list(lambda0, c, Q),
                              events, 
                              start = 0){
