+++ R/interpolate_mmhp_latent.R
@@ -7,8 +7,8 @@
 #' @return list of the states of the Markov process (z.hat), including 
 #' initial state
 #' and the times of the transitions between these times (x.hat).
-#' @noRd
-#'
+#' @keywords internal
+#' 
 #' @examples
 #' Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
 #' mmhp_obj <- pp_mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, lambda1 = 1.1,
