+++ R/pearsonresidual.R
@@ -8,10 +8,10 @@
 #' @param start start of observation period (default 0)
 #' @param end termination time (default final event)
 #' @param steps number of steps for numeric integration (if needed)
-#' 
 #' @return the Pearson residual
 #' @importFrom stats integrate
-#' @noRd
+#' @keywords internal
+#' @export
 #' @examples 
 #' Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
 #' x <- pp_mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, 
@@ -19,17 +19,20 @@
 #' y <- pp_simulate(x, n = 10)
 #' ppdiag:::pearsonresidual(x, events = y$events[-1])
 
+
 pearsonresidual <- function(object, events, start, end, steps = 1000) {
   UseMethod("pearsonresidual")
 }
 
-
+#' @keywords internal
+#' @export
 pearsonresidual.default <- function(object, events, start = 0,
                                 end = max(events), steps = 1000) {
   cat("Please input the right model. Select from hp, hpp, mmpp and mmhp. ")
 }
 
-
+#' @keywords internal
+#' @export
 pearsonresidual.mmhp <- function(object, events, start = 0,
                                 end = max(events), steps = 1000) {
   if(end != max(events)) {
@@ -54,7 +57,8 @@ pearsonresidual.mmhp <- function(object, events, start = 0,
   return(pr)
 }
 
-
+#' @keywords internal
+#' @export
 pearsonresidual.hp <- function(object, events, start = 0,
                                 end = max(events), steps = 1000) {
   if(end != max(events)) {
@@ -117,7 +121,8 @@ pearsonresidual.hp <- function(object, events, start = 0,
   }
 }
 
-
+#' @keywords internal
+#' @export
 pearsonresidual.hpp <- function(object, events, start = 0,
                                 end = max(events), steps = 1000) {
   if(end != max(events)) {
@@ -133,8 +138,8 @@ pearsonresidual.hpp <- function(object, events, start = 0,
   return(pr)
 }
 
-
-
+#' @keywords internal
+#' @export
 pearsonresidual.mmpp <- function(object, events, start = 0,
                                  end = max(events), steps = 1000) {
   if(end != max(events)) {
