+++ R/pp_compensator.R
@@ -1,16 +1,15 @@
 #' Compensators for point processes
-#' 
+#'
 #' Computes the compensator for included point processes
 #'
 #' @param object a point process model
 #' @param events event times, which can have first value as 0
 #' @return compensator vector of rescaled interevent times
 #' @export
-#' @examples 
+#' @examples
 #' hpp_obj <- pp_hpp(lambda = 1)
-#' events <- pp_simulate(hpp_obj, end=10)
+#' events <- pp_simulate(hpp_obj, end = 10)
 #' comp <- pp_compensator(hpp_obj, events)
-
 pp_compensator <- function(object, events) {
   UseMethod("pp_compensator")
 }
@@ -29,18 +28,18 @@ pp_compensator.mmpp <- function(object, events) {
   Q <- object$Q
   q1 <- Q[1, 2]
   q2 <- Q[2, 1]
-  if(events[1] == 0) {
+  if (events[1] == 0) {
     n <- length(events) - 1
     interevent <- events[-1] - events[-(n + 1)]
     pzt <- mmpp_event_state(params = object, events[-1])$pzt
   }
   else {
     n <- length(events)
-    interevent <- diff(c(0,events))
+    interevent <- diff(c(0, events))
     pzt <- mmpp_event_state(params = object, events)$pzt
   }
   interevent <- events[-1] - events[-(n + 1)]
-  Lambda_mixed <- lambda0 * (1 + c) * interevent * pzt + 
+  Lambda_mixed <- lambda0 * (1 + c) * interevent * pzt +
     lambda0 * interevent * (1 - pzt)
   return(Lambda_mixed)
 }
@@ -48,34 +47,33 @@ pp_compensator.mmpp <- function(object, events) {
 #' @rdname pp_compensator
 #' @export
 pp_compensator.hp <- function(object, events) {
-  # input object: parameters for Hawkes process, include lambda0, alpha, beta 
+  # input object: parameters for Hawkes process, include lambda0, alpha, beta
   #       events: vector of events times
   # output Lambda: vector of compensator evaluated at each event time
-  
+
   lambda0 <- object$lambda0
   alpha <- object$alpha
   beta <- object$beta
-  if(events[1] == 0 ){
+  if (events[1] == 0) {
     events <- events[-1]
   }
-  if(length(events) == 0) {
+  if (length(events) == 0) {
     return(0)
   }
   N <- length(events)
-  Lambda <- rep(0,N)
+  Lambda <- rep(0, N)
   r <- 0
-  Lambda[1] <- lambda0*(events[1]) # start at 0
-  if(N == 1) {
+  Lambda[1] <- lambda0 * (events[1]) # start at 0
+  if (N == 1) {
     return(Lambda)
   }
-  for(i in 2:N){
-    delta.t <- events[i]-events[i-1]
-    temp.r <- exp(-beta*delta.t)*(r+1)
-    Lambda[i] <- lambda0*delta.t-alpha/beta*(temp.r-r-1)
+  for (i in 2:N) {
+    delta.t <- events[i] - events[i - 1]
+    temp.r <- exp(-beta * delta.t) * (r + 1)
+    Lambda[i] <- lambda0 * delta.t - alpha / beta * (temp.r - r - 1)
     r <- temp.r
   }
   return(Lambda)
-
 }
 
 #' @rdname pp_compensator
@@ -87,27 +85,27 @@ pp_compensator.mmhp <- function(object, events) {
   beta <- object$beta
   q1 <- object$Q[1, 2]
   q2 <- object$Q[2, 1]
-  if(length(events) == 0){
+  if (length(events) == 0) {
     return(0)
   }
-  if(events[1] == 0) {
[…]
