+++ R/pp_diag.R
@@ -1,7 +1,7 @@
 #' Summarise diagnostics for point process models
 #'
-#' Generate diagnostic tools for different point process models, 
-#' including quantile-quantile plot, ks plot, 
+#' Generate diagnostic tools for different point process models,
+#' including quantile-quantile plot, ks plot,
 #' raw residual and pearson residual.
 #'
 #' @param object a point process model
@@ -12,11 +12,10 @@
 #' @return Display QQ-plot and KS plot, along with printing Pearson
 #'  and raw residuals.
 #' @export
-#' @examples 
+#' @examples
 #' hpp_obj <- pp_hpp(lambda = 1)
-#' events <- pp_simulate(hpp_obj, end=50)
+#' events <- pp_simulate(hpp_obj, end = 50)
 #' pp_diag(hpp_obj, events)
-
 pp_diag <- function(object, events) {
   UseMethod("pp_diag")
 }
@@ -29,86 +28,94 @@ pp_diag.default <- function(object, events) {
 
 #' @rdname pp_diag
 #' @export
-pp_diag.hp<-function(object, events){
+pp_diag.hp <- function(object, events) {
   r <- pp_compensator(object, events)
   oldpar <- par(no.readonly = TRUE)
   on.exit(par(oldpar))
-  layout(mat = matrix(c(1,2), nrow = 1, ncol = 2),
-         heights = c(2, 2),
-         widths = c(2, 2))
+  layout(
+    mat = matrix(c(1, 2), nrow = 1, ncol = 2),
+    heights = c(2, 2),
+    widths = c(2, 2)
+  )
   par(mar = c(2, 2, 1, 1))
   pp_qqexp(r)
   par(mar = c(2, 2, 1, 1))
   pp_ksplot(r)
   rr <- rawresidual(object, events, end = max(events))
   pr <- pearsonresidual(object, events, end = max(events))
-  ks <- ks.test(r,"pexp")
-  cat("Raw residual: ", rr, "\n",sep = "")
-  cat("Pearson residual: ", pr, "\n",sep = "")
+  ks <- ks.test(r, "pexp")
+  cat("Raw residual: ", rr, "\n", sep = "")
+  cat("Pearson residual: ", pr, "\n", sep = "")
   print(ks)
   return(NULL)
 }
 
 #' @rdname pp_diag
 #' @export
-pp_diag.mmhp <- function(object, events){
+pp_diag.mmhp <- function(object, events) {
   r <- pp_compensator(object, events)
   oldpar <- par(no.readonly = TRUE)
   on.exit(par(oldpar))
-  layout(mat = matrix(c(1,2), nrow = 1, ncol = 2), heights = c(2, 2),
-         widths = c(2, 2))
+  layout(
+    mat = matrix(c(1, 2), nrow = 1, ncol = 2), heights = c(2, 2),
+    widths = c(2, 2)
+  )
   par(mar = c(2, 2, 1, 1))
   pp_qqexp(r)
   par(mar = c(2, 2, 1, 1))
   pp_ksplot(r)
   rr <- rawresidual(object, events, end = max(events))
   pr <- pearsonresidual(object, events, end = max(events))
-  ks <- ks.test(r,"pexp")
-  cat("Raw residual: ", rr, "\n",sep = "")
-  cat("Pearson residual: ", pr, "\n",sep = "")
+  ks <- ks.test(r, "pexp")
+  cat("Raw residual: ", rr, "\n", sep = "")
+  cat("Pearson residual: ", pr, "\n", sep = "")
   print(ks)
   return(NULL)
 }
 
 #' @rdname pp_diag
 #' @export
[…]

[…]
