+++ CONTRIBUTING.md
@@ -1,0 +1,7 @@
+# Contributing to ppdiag development
+
+The goal of this guide is to help you set up and contribute to ppdiag as 
+quickly as possible. 
+
+Please note that ppdiag is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, 
+you agree to abide by its terms.
