### 2021-05-11 aa7dd83
Tweak citation (#55)
This is part of openjournals/joss-reviews#3133

Minor tweak to the double citation to match the preferred [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#Citation_Syntax).
