### 2021-05-12 d57fec3
update contributing.md and add generic print functions instead of a specific function.
