+++ man/pp_print.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pp_print.R
-\name{pp_print}
-\alias{pp_print}
-\title{Print temporal point process objects}
-\usage{
-pp_print(object)
-}
-\arguments{
-\item{object}{point process model object of type hpp, hp, mmhp, or mmpp}
-}
-\value{
-no return value, print object attributes
-}
-\description{
-Currently available point processes are homogeneous Poisson,
-Hawkes with exponential kernel, MMHP and MMPP
-}
-\examples{
-hpp_obj <- pp_hpp(lambda = 1)
-pp_print(hpp_obj)
-}
