+++ CONTRIBUTING.md
@@ -1,7 +1,18 @@
 # Contributing to ppdiag development
 
-The goal of this guide is to help you set up and contribute to ppdiag as 
-quickly as possible. 
+Bug reports and any potential problems with any existing code can be addressed
+by filing an issue giving some brief details and if possible
+a small reproducible example. If you would like to fix the bug 
+yourself and it is a small fix, feel free to go ahead with a direct
+pull request. Larger issues might be better proceeded by some discussion.
+
+We would be delighted for anyone who is interested in contributing to
+this package to get involved also. Potential additions would be to extend the
+included diagnostic tests to other point processes, such as different
+Hawkes kernel functions. 
+
+To contribute, please look at the existing code and feel free to file
+an issue to discuss further any potential additions.
 
 Please note that ppdiag is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, 
 you agree to abide by its terms.
