+++ NAMESPACE
@@ -25,6 +25,10 @@ S3method(pp_simulate,hp)
 S3method(pp_simulate,hpp)
 S3method(pp_simulate,mmhp)
 S3method(pp_simulate,mmpp)
+S3method(print,hp)
+S3method(print,hpp)
+S3method(print,mmhp)
+S3method(print,mmpp)
 S3method(rawresidual,default)
 S3method(rawresidual,hp)
 S3method(rawresidual,hpp)
@@ -45,7 +49,6 @@ export(pp_hpp)
 export(pp_ksplot)
 export(pp_mmhp)
 export(pp_mmpp)
-export(pp_print)
 export(pp_qqexp)
 export(pp_residual)
 export(pp_simulate)
