### 2021-05-12 ceb2498
joss-fix (see comment) (#56)
* add print method pp_print, fix a small typo for pp_simulate

* change @keywards Internal to internal, so no need to add @noRd

* change @keywards Internal to internal, also added return NULL for pp_diag

* get rid of auto-generated git amendment message

* clean up white space and fix indentations using stylish

* add code_of_conduct.md and contributing.md(contributing file is not done)

* update contributing.md and add generic print functions instead of a specific function.
