+++ R/pearsonresidual.R
@@ -1,6 +1,6 @@
 #' Compute Pearson residuals for point process models
 #'
-#' Compute Pearson residuals for point processes 
+#' Compute Pearson residuals for point processes
 #' with specified parameters and events.
 #'
 #' @param object point process model
@@ -8,34 +8,37 @@
 #' @param start start of observation period (default 0)
 #' @param end termination time (default final event)
 #' @param steps number of steps for numeric integration (if needed)
-#' 
 #' @return the Pearson residual
 #' @importFrom stats integrate
-#' @noRd
-#' @examples 
+#' @keywords internal
+#' @export
+#' @examples
 #' Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
-#' x <- pp_mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, 
-#' lambda1 = 1.1, alpha = 0.8, beta = 1.2)
+#' x <- pp_mmhp(Q,
+#'   delta = c(1 / 3, 2 / 3), lambda0 = 0.9,
+#'   lambda1 = 1.1, alpha = 0.8, beta = 1.2
+#' )
 #' y <- pp_simulate(x, n = 10)
 #' ppdiag:::pearsonresidual(x, events = y$events[-1])
-
 pearsonresidual <- function(object, events, start, end, steps = 1000) {
   UseMethod("pearsonresidual")
 }
 
-
+#' @keywords internal
+#' @export
 pearsonresidual.default <- function(object, events, start = 0,
-                                end = max(events), steps = 1000) {
+                                    end = max(events), steps = 1000) {
   cat("Please input the right model. Select from hp, hpp, mmpp and mmhp. ")
 }
 
-
+#' @keywords internal
+#' @export
 pearsonresidual.mmhp <- function(object, events, start = 0,
-                                end = max(events), steps = 1000) {
-  if(end != max(events)) {
+                                 end = max(events), steps = 1000) {
+  if (end != max(events)) {
     message("PR calculated to specified end time.")
   }
-  if(events[1] == 0 ) {
+  if (events[1] == 0) {
     events <- events[-1]
   }
   # define time.vec,latent.vec,latent_event in intensity
@@ -45,22 +48,27 @@ pearsonresidual.mmhp <- function(object, events, start = 0,
   event_obj$start <- start
   event_obj$end <- end
   time.vec <- seq(from = start, to = end, length.out = steps)
-  est.intensity <- pp_intensity(object, event_info = event_obj,
-                             method = "numeric", steps = steps)
-  est.intensity.events <- pp_intensity(object, event_info = event_obj,
-                                    method = "atevent")
+  est.intensity <- pp_intensity(object,
+    event_info = event_obj,
+    method = "numeric", steps = steps
+  )
+  est.intensity.events <- pp_intensity(object,
+    event_info = event_obj,
+    method = "atevent"
+  )
   pr <- sum(1 / sqrt(est.intensity.events)) -
     sum(sqrt(est.intensity)) * (time.vec[2] - time.vec[1])
   return(pr)
 }
 
-
+#' @keywords internal
+#' @export
 pearsonresidual.hp <- function(object, events, start = 0,
-                                end = max(events), steps = 1000) {
-  if(end != max(events)) {
+                               end = max(events), steps = 1000) {
+  if (end != max(events)) {
     message("PR calculated to specified end time.")
   }
-  if(events[1] == 0) {
+  if (events[1] == 0) {
     events <- events[-1]
   }
   lambda0 <- object$lambda0
@@ -78,8 +86,10 @@ pearsonresidual.hp <- function(object, events, start = 0,
     integrand <- function(u) {
[…]
