### 2021-05-12 cc0bebc
update to remove indiv residual functions from being exported
