+++ .github/workflows/pkgdown.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: Install dependencies
         run: |
           remotes::install_deps(dependencies = TRUE)
-          remotes::install_github("r-lib/pkgdown", dependencies = TRUE)
+          remotes::install_cran("pkgdown", dependencies = TRUE)
         shell: Rscript {0}
 
       - name: Install package
