### 2021-05-13 d786be7
update to paper making clearer the uses of point processes
