### 2021-05-19 b4c3d92
Joss paper (#57)
* add joss-paper

* Tweak citation (#55)

This is part of openjournals/joss-reviews#3133

Minor tweak to the double citation to match the preferred [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#Citation_Syntax).

* update to paper making clearer the uses of point processes

* fixes #58

Co-authored-by: Jake Thompson <wjakethompson@users.noreply.github.com>
Co-authored-by: Qian Sun <qiansun@brandeis.edu>
