+++ paper.md
@@ -1,166 +1,0 @@
----
-title: 'ppdiag: Diagnostic Tools for Temporal Point Processes'
-tags:
-  - R
-  - Diagnostics
-  - Goodness of fit
-  - Temporal Point Processes
-authors:
-  - name: Sally Sun
-    affiliation: 1 # (Multiple affiliations must be quoted)
-  - name: Owen G. Ward
-    orcid: 0000-0002-9678-3542
-    affiliation: 1
-  - name: Jing Wu
-    affiliation: 1
-  - name: Lihao Xiao
-    affiliation: 1
-  - name: Xiaoxi Zhao
-    affiliation: 1
-  - name: Tian Zheng
-    affiliation: "1,2"
-affiliations:
- - name: Department of Statistics, Columbia University
-   index: 1
- - name: Data Science Institute, Columbia University
-   index: 2
-citation_author: Sun et. al.
-date: "2021-05-13"
-year: 2021
-bibliography: paper.bib
-output: rticles::joss_article
-csl: apa.csl
-journal: JOSS
----
-
-
-
-# Summary
-
-Temporal Point Processes are widely used to model
-phenomena in many fields, such as finance and neuroscience.
-However, tools to evaluate the fit of these point
-processes to data, and to identify reasons for lack of
-fit, are not readily implemented in common software.
-Here we provide `ppdiag`, an `R`
-package containing a 
-selection of statistically motivated
-tools to analyse the goodness of fit of point processes to
-data, as have been described in @wu2020diagnostics.
-
-# Statement of Need
-Temporal point processes can be used to model data arising in a wide
-variety of fields, including but not limited to neuroscience, seismology
-and finance. As such, tools for analysing the fit of 
-point processes to data can be used by many practitioners.
-This package provides functions to evaluate the fit of
-univariate temporal point processes. These functions allow: 
-
-- Simulation 
-of data from a range of common univariate point processes,
-including Homogeneous 
-Poisson Process, Hawkes Process, and Markov-modulated Hawkes and Poisson
-Processes. 
-
-- Fitting common univariate point processes to data,
-and plotting the intensity function over data. 
-
-- After fitting a point process model to data, evaluating the 
-ability of that model to capture the temporal
-structure present in data.
-Methods for diagnostics include computing raw and Pearson residuals,
-a Kolmogrov-Smirnov 
-test (based on the time rescaling theorem)
-and corresponding diagnostic plots. These diagnostics can be used to 
-compare the fit of multiple point processes to data and to
-identify reasons for lack of fit of a process to data.
-
-
-Though there are some existing 
-packages that contain functions for simulating point 
-processes, [@hawkes_14],[@Markov_17],
-and for computing some 
-simple residuals [@PtProcess2010],
-to the best of our knowledge, 
-there are no existing packages that provide methods for
-diagnostics of different 
-temporal point processes and for comparing the fit of these point processes. 
-These diagnostic tools are useful to identify problems in model 
-fitting and understand the causes of this lack of fit.
-
-# Example
-
-We illustrate the use of this functionality
-with a simple example, using earthquake after-shock data
-[@PtProcess2010]. We consider all after shocks within a
-year of the original Phuket earthquake and wish to see 
-if they can be described by a temporal point process.
-
-
-```r
-set.seed(2021) 
-library(ppdiag)
-library(PtProcess)
-data("Phuket")
-end <- 365
-event_times <- Phuket$time[Phuket$time < end]
-```
-
[…]
