+++ paper.bib
@@ -28,6 +28,7 @@
     number = {8},
     pages = {1--32},
     url = {http://www.jstatsoft.org/v35/i08/},
+    doi={10.18637/jss.v035.i08}
 }
 
 @article{wu2020diagnostics,
