+++ paper.bib
@@ -31,11 +31,20 @@
     doi={10.18637/jss.v035.i08}
 }
 
-@article{wu2020diagnostics,
-  title={Diagnostics and Visualization of Point Process Models for Event Times on a Social Network},
-  author={Wu, Jing and Smith, Anna L and Zheng, Tian},
-  journal={arXiv preprint arXiv:2001.09359},
-  year={2020}
+@inbook{wu2021diagnostics,
+author = {Wu, Jing and Smith, Anna L. and  Zheng, Tian},
+publisher = {John Wiley & Sons, Ltd},
+isbn = {9781119821588},
+title = {Diagnostics and Visualization of Point Process Models for Event Times on a Social Network},
+booktitle = {Applied Modeling Techniques and Data Analysis 1},
+chapter = {7},
+pages = {129-145},
+doi = {https://doi.org/10.1002/9781119821588.ch7},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/9781119821588.ch7},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/9781119821588.ch7},
+year = {2021},
+keywords = {diagnostic statistics, network event times models, network heterogeneity, point process models, simulated studies, social network, time rescaling theorem, visualization tools},
+abstract = {Summary Point process models have been used to analyze interaction event times on a social network, in the hope to provide valuable insights for social science research. In this chapter, the authors aim to propose diagnostic statistics and visualization tools for network event times models, which they develop as extensions of evaluation techniques for univariate point process models. They investigate the proposed techniques using simulated studies and real data. The authors introduce the notation for point processes, network point processes and related models. They focus on time-domain diagnoses by applying the time rescaling theorem and inspecting residual processes. The diagnostic tools for detecting network heterogeneity and network structure in residual processes are developed and demonstrated.}
 }
 
 @Manual{hawkes_14,
