+++ paper.bib
@@ -39,7 +39,7 @@ title = {Diagnostics and Visualization of Point Process Models for Event Times o
 booktitle = {Applied Modeling Techniques and Data Analysis 1},
 chapter = {7},
 pages = {129-145},
-doi = {https://doi.org/10.1002/9781119821588.ch7},
+doi = {10.1002/9781119821588.ch7},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/9781119821588.ch7},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/9781119821588.ch7},
 year = {2021},
