### 2021-05-27 #3133 @arfon
@OwenWard – please merge this PR which makes some small changes to your paper (the metadata ones are essential for compatibility with our final publishing workflow steps): https://github.com/OwenWard/ppdiag/pull/59
