### 2021-05-27 2a71cd8
Making some small fixes to the paper. (#59)
